+++ src/Layers/xrRenderDX10/3DFluid/dx103DFluidEmitters.cpp
@@ -52,7 +52,7 @@ void dx103DFluidEmitters::DestroyShaders()
 void dx103DFluidEmitters::RenderDensity(const dx103DFluidData &FluidData)
 {
 	const xr_vector<CEmitter>	&Emitters = FluidData.GetEmittersList();
-	u32	iNumEmitters = Emitters.size();
+	u32	iNumEmitters = (u32)Emitters.size();
 
 	for( u32 i=0; i<iNumEmitters; ++i )
 	{
@@ -65,7 +65,7 @@ void dx103DFluidEmitters::RenderDensity(const dx103DFluidData &FluidData)
 void dx103DFluidEmitters::RenderVelocity(const dx103DFluidData &FluidData)
 {
 	const xr_vector<CEmitter>	&Emitters = FluidData.GetEmittersList();
-	u32	iNumEmitters = Emitters.size();
+	u32	iNumEmitters = (u32)Emitters.size();
 
 	for( u32 i=0; i<iNumEmitters; ++i )
 	{
